% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_PVE.R
\name{delta_PVE}
\alias{delta_PVE}
\title{Estimating the effective scale of neighbor effects}
\usage{
delta_PVE(res, fig = TRUE, ...)
}
\arguments{
\item{res}{Output results of \code{calc_PVEnei()}.}

\item{fig}{TRUE/FALSE to plot the results (or not). Default is TRUE.}

\item{...}{Arguments to be passed to \code{plot()}.}
}
\value{
Estimated effective scale and proportion of phenotypic variation explained by neighbor effects at that scale.
}
\description{
A function to calculate \eqn{\Delta}PVE that estimates the effective scale of neighbor effects.
}
\seealso{
calc_PVEnei
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
