% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{w}
\alias{w}
\title{Calculating a distance decay weight}
\usage{
w(s, a, kernel = "exp")
}
\arguments{
\item{s}{A numeric scalar indicating spatial distance at which the distance decay is referred}

\item{a}{A numeric scalar indicating a decay coefficient}

\item{kernel}{An option to select a negative exponential kernel \code{"exp"} or Gaussian kernel \code{"gaussian"}.}
}
\value{
A numeric scalar for a distance decay weight.
}
\description{
A function to calculate, with a negative exponential or Gaussian dispersal kernel.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
