% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_pheno_simu.R
\name{qtl_pheno_simu}
\alias{qtl_pheno_simu}
\title{Simulating phenotype values with neighbor effects.}
\usage{
qtl_pheno_simu(
  b_self,
  b_nei,
  eigenK_self,
  eigenK_nei,
  b_ratio = c(1, 1),
  pveB,
  pve
)
}
\arguments{
\item{b_self}{A n x 1 genotype vector to design major additive genetic effect.}

\item{b_nei}{A vector of explanatory variable for beta_2 (neighbor effects)}

\item{eigenK_self}{Products of function 'eigen' with self covariance matrices that were used as explanatory variables for the phenotype.}

\item{eigenK_nei}{Products of function 'eigen' with neighbor covariance matrices that were used as explanatory variables for the phenotype.}

\item{b_ratio}{Ratio for contribution of engenK_self,  eigenK_nei and eigenK_sxn to the phenotype.}

\item{pveB}{Proportion of variance explained by genetic effect designed by b_.. vector.}

\item{pve}{Proportion of variance explained by all genetic effects (i.e., b_.. and eigenK_..)}
}
\value{
A list of simulated phenotypes
\itemize{
\item{y}{Simulated phenotype values}
\item{beta_self}{major self-genetic effects}
\item{beta_nei}{major neighbor effects}
\item{sigma_self}{self polygenic effects}
\item{sigma_nei}{neighbor polygenic effects}
\item{epsilon}{residuals}
\item{res_pveB}{realized proportion of variation explained by major-effect genes}
\item{res_pve}{realized proportion of variation explained by major-effect genes and polygenic effects}
}
}
\description{
A function to simulate phenotype values with multiple sources of variation controlled
}
\author{
Eiji Yamamoto, and Yasuhiro Sato
}
